sched/walt: sysctl node to allow capture of walt issues
It is necessary to maintain debug checks in the walt code
while preventing errant crashes in released code.

Create a sysctl node that will control whether walt issues
will crash the system, and create a macro to take advantage
of it. Support the debug strings required by walt, so that
this can be easily incorporated into walt files.

Change-Id: Ie81cefe3c6d730574599f8d804d010f4a917bba7
Signed-off-by: Stephen Dickey <dickey@codeaurora.org>