UPSTREAM: sched: Fix DEBUG && !SCHEDSTATS warn
When !SCHEDSTATS schedstat_enabled() is an unconditional 0 and the
whole block doesn't exist, however GCC figures the scoped variable
'stats' is unused and complains about it.

Upgrade the warning from -Wunused-variable to -Wunused-but-set-variable
by writing it in two statements. This fixes the build because the new
warning is in W=1.

Given that whole if(0) {} thing, I don't feel motivated to change
things overly much and quite strongly feel this is the compiler being
daft.

Fixes: cb3e971c435d ("sched: Make struct sched_statistics independent of fair sched class")
Reported-by: Stephen Rothwell <sfr@canb.auug.org.au>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
(cherry picked from commit 769fdf83df57b373660343ef4270b3ada91ef434)

BUG=b:282905345
TEST=CQ

Disallow-Recycled-Builds: test-failures
Change-Id: I46faf4aa9ed3de10fcf3c8f71a3a7e91db52d78e
Signed-off-by: Aashay Shringarpure <aashay@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4638485
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: William K Lin <wklin@google.com>