ANDROID: arch_topology: Add android_rvh_update_thermal_stats
Add trace hook so that vendor modules can track
thermal changes across the system.

This feature is necessary to implement Qualcomm's
load tracking scheme, WALT. WALT was originally
submitted upstream, but due to complexities it was
decided that a better path forward would to be
improve on existing PELT. Link to upstream
submission of WALT can be found below.

Bug: 203763121
Link: https://lore.kernel.org/lkml/1477638642-17428-2-git-send-email-markivx@codeaurora.org/
Change-Id: I5c54b313d3c92c0514f18b07797dd20c9c6f5161
Signed-off-by: Shaleen Agrawal <shalagra@codeaurora.org>
Signed-off-by: Vamsi Krishna Lanka <quic_vamslank@quicinc.com>
Signed-off-by: Sai Harshini Nimmala <quic_snimmala@quicinc.com>