sched/walt: limited the dest_cpu with set_cpus_allowed_ptr_locked
When set_cpus_allowed_ptr_locked runs, it chooses a new destination
cpu for the task based upon the new mask for the task. Unfortunately
this means that halted cpus can be included.

Register for a trace hook such that dest_cpu can be recomputed, and
a dest_cpu that is not halted, chosen.

Change-Id: I0ab072eb395bece8c0ab896d7a1008f9cf55f468
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>