sched/walt/halt: update tracepoints for halt/start
halt/start tracepoints need a little enhancement, to ensure
a measurement can be made from the start of an operation.

Also to debug RT task placements, add the lowest_mask
cpu mask which contains the cpus available for the
RT task placement algorithm to trace_sched_cpu_util, and
update the trace to show the halted state.

Change-Id: Idd5ab03d8ca5af21d42415dce790993c8c5df1d5
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>