sched/walt: Fix excess detach_task in walt_lb_pull_tasks
There's three loops to find a task to pull.

In the first iteration, we look for a task that is not running,
which fits a strict criteria.

In the second iteration, we look for a task that is not running,
which fits a lax criteria.

If we have still not found a valid task to pull, we intended to pull the
running task servicing an active load migration.

However, there is some leftover code, wherein a task that is not running
will end up being detached and pulled. This was not intended, and leads
to incorrect behavior. Essentially, this is negating the effects of the
first two iterations.

Change-Id: Ia225d89043c76ec6e661771a39d058498346920d
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>