kernel/sched/walt: 64/32bit tasks can execve 32/64bit tasks
A task that is 64 bit can be come a 32bit task through execve,
ad the 32bit binary is loaded. Likewise a 32bit task can
become a 64bit task through execve.

Dec walt stats handles the 32 bit to 64 bit case, but doesn't
handle the 32 bit to 64bit case. Furthermore, when the task
is marked as misfit, a positive or negative change can take
placem in the misfit task accounting.

Generalize the adjustments made to account for misfit tasks,
and handle transisions either to or from 32bit tasks.

Change-Id: I6c746eaeb85301a325568f9f3c5c8d4a86c1e17b
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>