kernel/sched/walt: change formatting for not_preferred output
Core control takes a numeric input for the not_preferred field
to identify which cpus can be paused. 1 means can pause, 0 means
the cpu cannot pause.

The output is in order of the number of the cpu. For example

  # cat /sys/devices/system/cpu/cpu3/core_ctl/not_preferred
  1 0 1 0

This means that cpu 3 and cpu 5 are not_preferred and can be
paused.

Change-Id: Ia4b982352bba7e463d59579ece13ccfcb09b0baf
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>