sched/walt/init: Ignore Perf Domain for Single Cluster Systems
In the event that a cpu topology is a set of cpus with uniform
capacity, a single cluster system will be formed by walt. When
this happens there are no performance domains, and the root-domain's
perf-domain is empty.

This may cause Walt to behave differently, but it isn't in itself
a failure. Code attempting to ensure that a valid rd->pd exists
is therefore failing on these single cluster systems.

Update the code that catches when the rd->pd is invalid, such
that the failure is ignored for single cluster systems.

Change-Id: I06bea8f357c00ce96d91ec3833f3a5a6afa3e121
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>