sched/walt: when halt kicks an unhalted cpu, set overload flag
When halt kicks a recently unhalted cpu, and ensures that
walt_newidle_balance is called to pull a task. Since there
has been no recent opportunity to run rebalance_domains or
update_sd_lb_stats the rq->rd->overload flag is 0. When
walt_newidle_blance is called, the overload flag is checked
and no pull of a task from the busiest cpu will be attempted.

When the smp interface is used to call walt_newidle_balance, which
is only called from halt to kick the newly unhalted cpu, force
the overload flag check so that walt_newidle_balance can
function properly for both the halt and the normal newidle case.

Change-Id: Ib3f8372e2436aef5019e48a6b277b86a6fecd37c
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>