sched/walt/core_control: strict_nr_run calculation
With the introduction of cpu masks to determine the cpus
a particular cluster should consider for nrrun, the calculation
for strict_nr_run changed incorrectly. nr_strict_need must
always include the nr_scaled value for the min-capacity-cluster,
while including the overflow nr_scaled value for other clusters.

Current code will not include min-capacity-cluster strict need
because of the way the active cpus are counted.

Restore the original code for compute_cluster_nr_strict_need(),
as this code is a clean implementation of the desired
functionality.

Change-Id: Ia284a9b8506d613b473e6809231f3c176275e155
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>