sched/walt: Prevent cpu rejection if no unhalted undying cpus
When a cpu is dying, it may still be active. However if the
task is a kthread the dying cpu will be rejected by
is_cpu_allowed, so it is effectively not available.

android_rvh_is_cpu_allowed returns false if the cpu being tested
is halted, but there is one remaining active cpu in the system.
This means all cpus in the system will be rejected, because they
are either dying or halted.

Update the android_rvh_is_cpu_allowed hook to first check whether
there are sufficient active/unhalted cpus for userspace threads,
and then check if there are sufficient active/unhalted,undying
cpus for kthreads.

Change-Id: I79ebdeb400360098e247afd1b5be9d29c9e4d6ce
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>