sched/walt: do not include partial paused cpus in accounting
Partial paused cpus can run some tasks, but the outcome of running
those tasks should not affect which cpus in this cluster are
to be paused or unpaused.

Change the accounting to ignore the partially paused cpus when
examining this cluster, to ensure they contribute nothing (no nr_run,
no max_nr, etc).

Change-Id: I5f849e87bc1511c00e0c912185f9da3c236f0218
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>