sched/walt/halt: ensure per-cpu-kthreads affinity is restored
When a cpu is offlined it's per-cpu-kthreads get affined to
all cpus. When it is onlined, the per-cpu-kthreads are restored
to having a single cpu in their affinity mask.

If the cpu is onlined when it is both offline and halted,
android_rvh_set_cpus_allowed_by_task will reject the cpu, causing
__set_cpus_allowed_ptr_locked to reject the update of the affinity.
This means per-cpu-kthreads can get stuck with 0xff affinity.

Correct the situation for any kthread by ensuring that if
a task's affinity is so restricted as to have only a single
cpu in it, allow that cpu.

Additionally, ensure that android_rvh_set_cpus_allowed_by_task will
disallow any restriction of a 32bit thread in execve. There are
many things that a task can do on a 64bit cpu prior to actual
execution of the task. Ensure that recently-forked tasks can
continue to run on any 64bit cpu in the system while still in execve.

Change-Id: I3a1f580499d28be19136c17dd1c545dbe93eb769
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>