sched/walt: compute_cluster_nr_misfit_assist should sum nr_misfit
compute_cluster_nr_misfit_assist() was written as part of the
replacement for the nr_prev_assist calculation performed in older
code. The component this replaces was summing nr_misift calculation
for the prev-prev-cluster (silvers). The current code does not,
and instead sums nr, which will give it an excessive count as compared
to previous kernels.

Update compute_cluster_nr_misfit_assist() to sum nr_misfit, rather
than nr.

Change-Id: I8a359f93bb5d52c2444bbcba6d8bd4fbf71d7fc1
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>