sched/walt: getaffinity should not restrict based upon halt
A task's affinity should not be dependent on halt, because
a halted cpu will be available to be run on, sometimes. Various
benchmarks and test cases will use getaffinity to determine
what is available to the task to start with. If halted cpus are
not returned as valid cpus, these tasks will be resticted to
never running on those cpus. This is not ideal behavior.

Remove the android_rvh_sched_getaffinity tracehook, as the only
thing it was doing was restricting the affinity returned.

The corresponding android_rvh_sched_setaffinity is already removed
in the baseline.  None-the-less, if it is every re-inroduced it
should likewise not restrict the affinity of a task from including
halted cpus.

Change-Id: I23257f02a904e6440e897c121fae0c4faaf83090
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>