sched/walt: core control nr_busy calculation
The nr_busy calculation was being double accounted for titaniums,
since it re-ran the loop to count nr_busy, for each cpu in the
cluster. The original design of eval_need was to perform this
operation once per cluster, using the lru mechanism to traverse
the cpus in the list exactly once. During refactoring, the
extra iteration was included.

Restore the original intention behind this code from eval_need.
Traverse the lru list exactly once, one time for each cpu in the
cluster.

Change-Id: If97ccbba840827cdfd3a874d4f9cfe85f67a0999
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>