sched/walt: Count misfits for partially halted cpus
The current code will ignore misfits if it is on a
partially halted cpu. A partially halted cpu may
be helping with small tasks, but new work load which
could ultimately be a big task can end up on those cpus.
When that happens the task can stay there indefinitely,
even as the task reaches fmax and becomes misfit.

If cpus that are monitoring cpus for misfits can count
those misfits even on partially halted cpus, they can
unhalt early, and pull the misfit load on the partially
halted cpu.

Change compute_cluster_nr_misfit such that partially halted
cpus can be monitored and action on other clusters can be
taken.

Change-Id: Iab41fa60a66a90eae9035c0ee6ded0a657442937
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>