sched/walt: Add trace member to show sync state
Currently the only way to know for certain if system is in frequency
sync state is to look at the number of cpus in the partial halt mask
under the halt_cpus tracepoint.

Add a sync_state member in sched_task_util tracepoint to make it easier
to identify state when debugging.

Change-Id: I5fecd5f66a4064e344e8f767154fce4937996cf9
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>