sched/walt: pipeline_set_boost must handle auto and manual pipeline
pipeline_set_boost does not handle the case properly, where both
manual and auto pipeline hinting are enabled or disabled, independently.
For example, this series of events

 pipeline_set_boost(true, MANUAL_PIPELINE);
  - auto pipeline enabled
 pipeline_set_boost(true, AUTO_PIPELINE);
  - auto pipeline disabled\
 pipeline_set_boost(false, MANUAL_PIPELINE);

With the above, pipeline boost will be enabled, even though manual
pipeline is no longer requesting it be enabled, and auto pipeline
is disabled.

Correct the code to independently track state of the AUTO or MANUAL
pipeline, and using that information to decide if boost is currently
requested or not.

Change-Id: Ia31eb8cd45b417f55f1e6827953073f708820ce6
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>