sched/walt: Introduce WALT_BUG if mark start progresses too far
A crash was observed where a task's mark start was observed to be
greater than 1 window boundary away from the task's window start. This
is problematic as it indicates there was a UTRA call that took place,
where the runqueue's window start was updated, but the task's window
start wasn't.

Introduce bug_on to catch this issue earlier, and gain insights from the
stack trace and event to better understand the background.

Change-Id: Ia1546e537ea5722a6065581898e1516274ed53ea
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>
Signed-off-by: Srinivasarao Pathipati <quic_c_spathi@quicinc.com>