sched/walt: Update capacities instantly on capping
Currently, there is an up to a window's worth of delay between when
frequencies are capped/uncapped to when the capacities are updated to
reflect this capping.

While the change in the frequency itself may be delayed until the next
cpufreq update, it is prudent to update the cpu capacities to reflect
the latest capped/uncapped reality as soon as the settings take place as
rq locks are not required to update cpu capacties.

Change-Id: I93963ac04f7c5bf746d294c5bae586d58390b782
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>