cpuidle: Disable deeper cpu sleep state when display is on in qcom-lpm governor.
qcom-lpm governor has a param sleep_disabled, which, when active, improves both performance and energy efficiency (unexpectedly, but it was tested) - thanks dereference23 for a hint. But deeper sleep state is still useful when phone screen is off, so let's make it depend on the screen state. This solution is, in theory, is pretty universal for Qualcomm devices, as long as they use qcom-lpm and sde_kms.